requirements.txt
requests==2.32.5
python-dotenv==1.1.1
urllib3==2.5.0
pyyaml>=6.0